Finally got slewing working

+++ cpls/include/motors.hpp
@@ -6,7 +6,7 @@ struct motor_t {
   char inverted;
   int power;
   float scale;
-  unsigned long slewRate;
+  float slewRate;
   unsigned long tlast;
   void set(int power);
 };
@@ -14,7 +14,7 @@ struct motor_t {
 namespace motors {
 void set(motor_t motor, int power);
 int get(motor_t motor);
-motor_t init(unsigned char port, int inverted, long slewRate, float scale);
+motor_t init(unsigned char port, int inverted, float slewRate, float scale);
 namespace slew {
 extern motor_t list[11];
 extern TaskHandle handle;
